+++ endian.h
@@ -31,30 +31,30 @@
 // Little endian's byte buffer to host's endian
 // Host's endian to Little endian's byte buffer
 #ifdef LITTLE_ENDIAN
-#define H2LS(s)		(unsigned short)s
-#define H2LL(l)		(unsigned long)l
-#define L2HS(s)		(unsigned short)s
-#define L2HL(l)		(unsigned long)l
-#define H2LBS(b,s)	memcpy(&b,&s,sizeof(unsigned short))
-#define H2LBL(b,l)	memcpy(&b,&l,sizeof(unsigned long))
-#define LB2HS(s,b)	memcpy(&s,&b,sizeof(unsigned short))
-#define LB2HL(l,b)	memcpy(&l,&b,sizeof(unsigned long))
+#define H2LS(s)		(uint16_t)s
+#define H2LL(l)		(uint32_t)l
+#define L2HS(s)		(uint16_t)s
+#define L2HL(l)		(uint32_t)l
+#define H2LBS(b,s)	memcpy(&b,&s,sizeof(uint16_t))
+#define H2LBL(b,l)	memcpy(&b,&l,sizeof(uint32_t))
+#define LB2HS(s,b)	memcpy(&s,&b,sizeof(uint16_t))
+#define LB2HL(l,b)	memcpy(&l,&b,sizeof(uint32_t))
 #endif
 #ifdef BIG_ENDIAN
-#define H2LS(s)		(((unsigned short)s&0x00FF) << 8) | (((unsigned short)s&0xFF00) >> 8)
-#define H2LL(l)		(((unsigned long)l&0x000000FF) << 24) |\
-					(((unsigned long)l&0x0000FF00) <<8)|\
-					(((unsigned long)l&0x00FF0000) >> 8) |\
-					(((unsigned long)l&0xFF000000) >> 24) 
-#define L2HBS(s) 	(((unsigned short)s&0x00FF) << 8) | (((unsigned short)s&0xFF00) >> 8)
-#define L2HBS(l)		(((unsigned long)l&0x000000FF) << 24) |\
-					(((unsigned long)l&0x0000FF00) <<8)|\
-					(((unsigned long)l&0x00FF0000) >> 8) |\
-					(((unsigned long)l&0xFF000000) >> 24) 
-#define H2LBS(b,s)	memcpy(&b,&H2LS(s),sizeof(unsigned short))
-#define H2LBL(b,l)	memcpy(&b,&H2LL(l),sizeof(unsigned long))
-#define LB2HS(s,b)	memcpy(&s,&L2HS(s),sizeof(unsigned short))
-#define LB2HL(l,b)	memcpy(&l,&L2HL(l),sizeof(unsigned long))
+#define H2LS(s)		(((uint16_t)s&0x00FF) << 8) | (((uint16_t)s&0xFF00) >> 8)
+#define H2LL(l)		(((uint32_t)l&0x000000FF) << 24) |\
+					(((uint32_t)l&0x0000FF00) <<8)|\
+					(((uint32_t)l&0x00FF0000) >> 8) |\
+					(((uint32_t)l&0xFF000000) >> 24) 
+#define L2HBS(s) 	(((uint16_t)s&0x00FF) << 8) | (((uint16_t)s&0xFF00) >> 8)
+#define L2HBS(l)		(((uint32_t)l&0x000000FF) << 24) |\
+					(((uint32_t)l&0x0000FF00) <<8)|\
+					(((uint32_t)l&0x00FF0000) >> 8) |\
+					(((uint32_t)l&0xFF000000) >> 24) 
+#define H2LBS(b,s)	memcpy(&b,&H2LS(s),sizeof(uint16_t))
+#define H2LBL(b,l)	memcpy(&b,&H2LL(l),sizeof(uint32_t))
+#define LB2HS(s,b)	memcpy(&s,&L2HS(s),sizeof(uint16_t))
+#define LB2HL(l,b)	memcpy(&l,&L2HL(l),sizeof(uint32_t))
 #endif
 
 #endif
